feat(core): implicitDependencies accept glob patterns, not just `*`
Note: this change requires nx to properly handle implicitDependencies represented as lists of projects.
Although [documentation](https://github.com/nrwl/nx/blob/master/docs/shared/affected.md) promises such feature,
this doesn't seem to be implemented as of version 12.0.0.
This change was not tested. It was done in hope that once the implicitDependencies are handled correctly,
the globing on project names will work as well.